Fixed ACU Bug
Fixed the bug where ACU would remain in projects when idle and getting paused. Added logic to skip it when it focuses itself (hopefully doesnt cause problems elsewhere). It will not be added to project list and unpaused, when it is supposedly idle. Does also apply to SACU, so far no other known units this applies to.